# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FieldsBody": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          }
        },
        "title": "FieldsBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadBody": {
        "properties": {
          "cid": {
            "default": "",
            "title": "Cid",
            "type": "string"
          },
          "content_type": {
            "default": "application/octet-stream",
            "title": "Content Type",
            "type": "string"
          },
          "ctitle": {
            "default": "",
            "title": "Ctitle",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "doc_id": {
            "title": "Doc Id",
            "type": "string"
          },
          "doc_no": {
            "default": "",
            "title": "Doc No",
            "type": "string"
          },
          "file_type": {
            "default": "",
            "title": "File Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          }
        },
        "required": [
          "doc_id",
          "filename"
        ],
        "title": "UploadBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowBody": {
        "properties": {
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          }
        },
        "title": "WorkflowBody",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Browser-independent read-only facade for the authorized CTF target. Write operations are preview-only and never sent upstream. Local Bearer tokens never include upstream .iPSA cookie values.",
    "title": "iSStech Purchase Requisition Replay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/previews/attachments/upload": {
      "post": {
        "operationId": "preview_upload_v1_previews_attachments_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Upload V1 Previews Attachments Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Upload",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/create": {
      "post": {
        "operationId": "preview_create_v1_previews_purchase_requisitions_create_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Create V1 Previews Purchase Requisitions Create Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Create",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/{requisition_id}/adjust": {
      "post": {
        "operationId": "preview_adjust_v1_previews_purchase_requisitions__requisition_id__adjust_post",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Adjust V1 Previews Purchase Requisitions  Requisition Id  Adjust Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Adjust",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/{requisition_id}/approve": {
      "post": {
        "operationId": "preview_approve_v1_previews_purchase_requisitions__requisition_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Approve V1 Previews Purchase Requisitions  Requisition Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Approve",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/{requisition_id}/delete": {
      "post": {
        "operationId": "preview_delete_v1_previews_purchase_requisitions__requisition_id__delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Delete V1 Previews Purchase Requisitions  Requisition Id  Delete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Delete",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/{requisition_id}/edit": {
      "post": {
        "operationId": "preview_edit_v1_previews_purchase_requisitions__requisition_id__edit_post",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Edit V1 Previews Purchase Requisitions  Requisition Id  Edit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Edit",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/{requisition_id}/revoke": {
      "post": {
        "operationId": "preview_revoke_v1_previews_purchase_requisitions__requisition_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Revoke V1 Previews Purchase Requisitions  Requisition Id  Revoke Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Revoke",
        "tags": [
          "previews"
        ]
      }
    },
    "/v1/previews/purchase-requisitions/{requisition_id}/submit": {
      "post": {
        "operationId": "preview_submit_v1_previews_purchase_requisitions__requisition_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Preview Submit V1 Previews Purchase Requisitions  Requisition Id  Submit Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Submit",
        "tags": [
          "previews"
        ]
      }
    },
...
  }
}